Integrate CSRForm into CSRModal
