--- frontend/pages/admin/IntegrationsPage/cards/Mdm/components/RequestCSRModal/RequestCSRForm/_styles.scss
@@ -1,13 +1,0 @@
-.request-csr {
-  &__form {
-    .bottom-label {
-      .form-field {
-        margin-bottom: 2px;
-      }
-
-      p {
-        margin-top: 0px;
-      }
-    }
-  }
-}
