--- frontend/pages/admin/IntegrationsPage/cards/Mdm/components/RequestCSRModal/RequestCSRForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./RequestCSRForm";
