+++ frontend/pages/admin/IntegrationsPage/cards/Mdm/components/RequestCSRModal/_styles.scss
@@ -2,4 +2,16 @@
   &__info-header {
     margin-bottom: $pad-xlarge;
   }
+
+  &__form {
+    .bottom-label {
+      .form-field {
+        margin-bottom: 2px;
+      }
+
+      p {
+        margin-top: 0px;
+      }
+    }
+  }
 }
